fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
pymongo
pymilvus
aiohttp
python-dotenv
